requirement.txt
asgiref==3.6.0
dj-database-url==2.0.0
Django==4.2.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==18.9.0
Pillow==9.3.0
psycopg2==2.9.3
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
